COMP: Add explicit casts at MSVC C4244 narrowing sites
MSVC C4244 (possible loss of data) fires at these sites once the blanket
4244 suppression is removed; make each narrowing explicit with
static_cast. Behavior is unchanged.

+++ Modules/Core/Common/include/itkProcessObject.h
@@ -924,7 +924,7 @@ protected:
   static inline constexpr float
   progressFixedToFloat(uint32_t fixed)
   {
-    return static_cast<double>(fixed) / static_cast<double>(std::numeric_limits<uint32_t>::max());
+    return static_cast<float>(static_cast<double>(fixed) / static_cast<double>(std::numeric_limits<uint32_t>::max()));
   }
 
   /**
